+++ lib/CMakeLists.txt
@@ -6,6 +6,5 @@ add_subdirectory(posix)
 endif()
 add_subdirectory_ifdef(CONFIG_CPP                  cpp)
 add_subdirectory(os)
-add_subdirectory(util)
 add_subdirectory_ifdef(CONFIG_SMF smf)
 add_subdirectory_ifdef(CONFIG_OPENAMP              open-amp)
