+++ lib/Kconfig
@@ -13,8 +13,6 @@ source "lib/posix/Kconfig"
 
 source "lib/open-amp/Kconfig"
 
-source "lib/util/Kconfig"
-
 source "lib/smf/Kconfig"
 
 endmenu
