--- lib/util/CMakeLists.txt
@@ -1,3 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-add_subdirectory_ifdef(CONFIG_FNMATCH		fnmatch)
