--- lib/util/Kconfig
@@ -1,9 +1,0 @@
-# Copyright (c) 2021 Nordic Semiconductor ASA
-# SPDX-License-Identifier: Apache-2.0
-
-
-menu "Util libraries"
-
-source "lib/util/fnmatch/Kconfig"
-
-endmenu
